+++ CHANGELOG.md
@@ -11,3 +11,11 @@
 # 2.2.1
 
 - Fix an issue when `queryStringToObject` function was receiving an empty string
+
+# X.X.X
+
+### Features
+
+##### Progress
+- `linesPerSecond` and `remainingTime` are now respectively named `linesPerSecondAvg` and `remainingTimeAvg`.
+- `linesPerSecond` and `remainingTime` now contains the value related to the previous computed snapshot.